Move debug screen to its own file, layout improvements, copy json
