+++ android/app/src/main/java/com/fleetdm/agent/MainActivity.kt
@@ -155,123 +155,6 @@ fun MainScreen(onNavigateToDebug: () -> Unit) {
 }
 
 @Composable
-fun DebugScreen(onNavigateBack: () -> Unit) {
-    val context = LocalContext.current
-
-    val restrictionsManager = context.getSystemService(RestrictionsManager::class.java)
-        ?: error("RestrictionsManager not available")
-    val appRestrictions = restrictionsManager.applicationRestrictions
-    val dpm = context.getSystemService(DevicePolicyManager::class.java)
-        ?: error("DevicePolicyManager not available")
-
-    val orchestrator = remember { AgentApplication.getCertificateOrchestrator(context) }
-    val delegatedScopes = remember { dpm.getDelegatedScopes(null, context.packageName).toList() }
-    val enrollmentSpecificID = remember { appRestrictions.getString("host_uuid")?.let { "****" + it.takeLast(4) } }
-    val hostCertificates = remember { orchestrator.getHostCertificates(context) }
-    val permissionsList = remember {
-        val packageInfo = context.packageManager.getPackageInfo(context.packageName, PackageManager.GET_PERMISSIONS)
-        val permissions = packageInfo.requestedPermissions ?: return@remember emptyList()
-        val flags = packageInfo.requestedPermissionsFlags ?: return@remember emptyList()
-
-        permissions.zip(flags.toList())
-            .filter { (_, flag) -> flag and PackageInfo.REQUESTED_PERMISSION_GRANTED != 0 }
-            .map { (permission, _) -> permission }
-    }
-    val fleetBaseUrl = remember { appRestrictions.getString("server_url") }
-    val baseUrl by ApiClient.baseUrlFlow.collectAsStateWithLifecycle(initialValue = null)
-    val installedCerts by orchestrator.installedCertsFlow(context).collectAsStateWithLifecycle(initialValue = emptyMap())
-
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        topBar = {
-            TopAppBar(
-                title = { Text("Debug Information") },
-                navigationIcon = {
-                    IconButton(onClick = onNavigateBack) {
-                        Icon(
-                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = "Back",
-                        )
-                    }
-                },
-            )
-        },
-        content = { paddingValues ->
-            Column(
-                Modifier
-                    .padding(paddingValues = paddingValues)
-                    .verticalScroll(rememberScrollState()),
-            ) {
-                KeyValue("packageName", context.packageName)
-                KeyValue("versionName", BuildConfig.VERSION_NAME)
-                KeyValue("longVersionCode", BuildConfig.VERSION_CODE.toString())
-                KeyValue("delegatedScopes", delegatedScopes.toString())
-                KeyValue("host_uuid (MC)", enrollmentSpecificID)
-                KeyValue("server_url (MC)", fleetBaseUrl)
-                KeyValue("server_url (DS)", baseUrl)
-                KeyValue("host_certificates", hostCertificates?.map { "${it.id}:${it.operation}" }.toString())
-                DebugCertificateList(certificates = installedCerts)
-                PermissionList(
-                    permissionsList = permissionsList,
-                )
-            }
-        },
-    )
-}
-
-@Composable
-fun DebugCertificateList(certificates: CertificateStateMap) {
-    Column {
-        Text("certificate status:", fontWeight = FontWeight.Bold)
-        certificates.forEach { (key, value) ->
-            Row(modifier = Modifier.padding(bottom = 5.dp, start = 10.dp)) {
-                Text(
-                    text = key.toString(),
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier.padding(end = 5.dp),
-                )
-                Column {
-                    Text(text = "alias: ${value.alias}")
-                    Text(text = "status: ${value.status}")
-                    Text(text = "retries: ${value.retries}")
-                    Text(text = "uuid: ${value.uuid}")
-                    Text(text = "notBefore: ${value.notBefore}")
-                    Text(text = "notAfter: ${value.notAfter}")
-                    Text(text = "serial: ${value.serialNumber}")
-                }
-            }
-        }
-        HorizontalDivider()
-    }
-}
-
-@Composable
-fun PermissionList(modifier: Modifier = Modifier, permissionsList: List<String>) {
-    Column(modifier = modifier) {
-        Text(text = "permission list:", fontWeight = FontWeight.Bold)
-        permissionsList.forEach {
-            Row {
-                Text(text = "- ", modifier = Modifier.padding(end = 8.dp))
-                Text(text = it)
-            }
-        }
-    }
-}
-
-@Composable
-fun KeyValue(key: String, value: String?) {
-    Text(
-        buildAnnotatedString {
-            withStyle(style = SpanStyle(fontWeight = FontWeight.Bold)) {
-                append(key)
-            }
-            append(": $value")
-        },
-    )
-    HorizontalDivider()
-}
-
-@Composable
 fun AboutFleet(modifier: Modifier = Modifier, onLearnClick: () -> Unit = {}) {
     Column(modifier = modifier.padding(20.dp)) {
         Text(
@@ -360,19 +243,6 @@ fun FleetScreenPreview() {
             )
             AppVersion(onClick = {})
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DebugCertificateListPreview() {
-    MyApplicationTheme {
-        DebugCertificateList(
-            certificates = mapOf(
-                1 to CertificateState(alias = "WIFI-1", status = CertificateStatus.INSTALLED),
-                2 to CertificateState(alias = "VPN-3", status = CertificateStatus.FAILED),
-            ),
-        )
     }
 }
